requirements.txt
opencv-python
numpy
matplotlib
tensorflow
scikit-learn
pillow
